reworked component structure
