+++ client/src/components/my-mentor-card/index.tsx
@@ -6,13 +6,11 @@ The full terms of this copyright and license should always be found in the root 
 */
 import React from "react";
 import {
-  Box,
   Card,
   CardContent,
   Typography,
   Tooltip,
   Avatar,
-  CircularProgress,
   Grid,
 } from "@material-ui/core";
 import StageToast from "./stage-toast";
@@ -22,57 +20,8 @@ import { Mentor } from "types";
 import { useWithThumbnail } from "hooks/graphql/use-with-thumbnail";
 import RecommendedActionButton from "./recommended-action-button";
 import { toTitleCase } from "helpers";
-
-function StageProgress(props: { value: number; max: number; percent: number }) {
-  return (
-    <Box alignItems="center">
-      <Box position="relative" display="inline-flex">
-        <CircularProgress
-          data-cy="stage-progress"
-          variant="determinate"
-          style={{ color: "lightgrey" }}
-          value={100}
-          size={80}
-        />
-
-        <Box
-          top={0}
-          left={0}
-          bottom={0}
-          right={0}
-          position="absolute"
-          display="flex"
-          alignItems="center"
-          justifyContent="center"
-        >
-          <CircularProgress
-            data-cy="stage-progress"
-            variant="determinate"
-            value={props.percent}
-            size={80}
-          />
-        </Box>
-        <Box
-          top={0}
-          left={0}
-          bottom={0}
-          right={0}
-          position="absolute"
-          display="flex"
-          alignItems="center"
-          justifyContent="center"
-        >
-          <Typography variant="h5" component="div" color="textSecondary">
-            {props.value}/{props.max}
-          </Typography>
-        </Box>
-      </Box>
-      <Typography variant="body1" color="textSecondary">
-        {props.percent}%
-      </Typography>
-    </Box>
-  );
-}
+import StageProgress from "./stage-progress";
+import parseMentor from "./mentor-info";
 
 const useStyles = makeStyles(() => ({
   homeThumbnail: {
@@ -84,95 +33,15 @@ const useStyles = makeStyles(() => ({
     height: 135,
   },
 }));
-const StageSelect = (value: number) => {
-  const stages = [
-    {
-      name: "Incomplete",
-      index: 0,
-      description: "This Mentor can't be built yet.",
-      floor: 0,
-      max: 5,
-    },
-    {
-      name: "Scripted",
-      index: 1,
-      description: "This Mentor can select questions from a list",
-      floor: 5,
-      max: 20,
-    },
-    {
-      name: "Interactive",
-      index: 2,
-      description: "This Mentor can respond to simple questions.",
-      floor: 20,
-      max: 50,
-    },
-    {
-      name: "Specialist",
-      index: 3,
-      description: "This mentor can answer questions within a specific topic.",
-      floor: 50,
-      max: 150,
-    },
-    {
-      name: "Conversational",
-      index: 4,
-      description: "This mentor can respond to questions with some nuance.",
-      floor: 150,
-      max: 250,
-    },
-    {
-      name: "Full-Subject",
-      index: 5,
-      description:
-        "Your mentor is equipped to answer questions within a broad subject.",
-      floor: 250,
-      max: 1000,
-    },
-    {
-      name: "Life-Story",
-      index: 6,
-      description:
-        "Your mentor can hold a natural conversation. Congratulations!",
-      floor: 1000,
-      max: value + 1,
-    },
-    {
-      name: "None",
-      index: 7,
-      description: "you've reached the final stage",
-      floor: value,
-      max: value,
-    },
-  ];
-  let currentStage = stages.find((stage) => {
-    return stage.max - 1 >= value;
-  });
-  if (!currentStage) currentStage = stages[0];
-  return {
-    ...currentStage,
-    ...{
-      next: stages[currentStage.index + 1],
-      percent: Math.round((value / currentStage.max) * 100),
-    },
-  };
-};
+
 export default function MyMentorCard(props: {
   accessToken: string;
   mentor: Mentor;
 }): JSX.Element {
-  const mentorId = props.mentor?._id || "";
-  const name = props.mentor?.name || "Unnamed";
-  const type = props.mentor?.mentorType;
-  const title = props.mentor?.title || "none";
-  const lastTrainedAt = props.mentor?.lastTrainedAt || "never";
-  const value =
-    props.mentor?.answers.filter((a) => a.status === "COMPLETE").length || 0;
-
-  const currentStage = StageSelect(value);
+  const mentorInfo = parseMentor(props.mentor);
   const classes = useStyles();
   const [thumbnail, updateThumbnail] = useWithThumbnail(
-    mentorId,
+    props.mentor?._id,
     props.accessToken,
     props.mentor?.thumbnail || ""
   );
@@ -195,14 +64,14 @@ export default function MyMentorCard(props: {
                 color="textSecondary"
                 data-cy="mentor-card-name"
               >
-                {name}
+                {mentorInfo.name}
               </Typography>
               <Typography
                 variant="h5"
                 color="textSecondary"
                 data-cy="mentor-card-info"
               >
-                Title: {title}
+                Title: {mentorInfo.title}
               </Typography>
               <Grid
                 justify="center"
@@ -244,7 +113,7 @@ export default function MyMentorCard(props: {
                 align="left"
                 data-cy="mentor-card-scope"
               >
-                Scope: {currentStage.name}
+                Scope: {mentorInfo.currentStage.name}
               </Typography>
               <Typography
                 variant="body1"
@@ -252,16 +121,16 @@ export default function MyMentorCard(props: {
                 align="left"
                 data-cy="mentor-card-scope-description"
               >
-                {currentStage.description}
+                {mentorInfo.currentStage.description}
               </Typography>
-              {type ? (
+              {mentorInfo.type ? (
                 <Typography
                   variant="h6"
                   color="textSecondary"
                   align="left"
                   data-cy="mentor-card-type"
                 >
-                  {toTitleCase(type)} Mentor
+                  {toTitleCase(mentorInfo.type)} Mentor
                 </Typography>
               ) : (
                 <Typography
@@ -280,20 +149,20 @@ export default function MyMentorCard(props: {
                 align="left"
                 data-cy="mentor-card-trained"
               >
-                Last Trained: {lastTrainedAt.substring(0, 10)}
+                Last Trained: {mentorInfo.lastTrainedAt.substring(0, 10)}
               </Typography>
             </Grid>
             <Grid item alignItems="center" xs={12} md={3}>
               <Typography variant="body1" color="textSecondary">
-                Next Goal: {currentStage.next.name}
+                Next Goal: {mentorInfo.currentStage.next.name}
                 {"   "}
                 <Tooltip
                   title={
                     <React.Fragment>
                       <Typography color="inherit">
-                        {currentStage.next.name}
+                        {mentorInfo.currentStage.next.name}
                       </Typography>
-                      {currentStage.next.description}
+                      {mentorInfo.currentStage.next.description}
                     </React.Fragment>
                   }
                   data-cy="next-stage-info"
@@ -302,18 +171,17 @@ export default function MyMentorCard(props: {
                 </Tooltip>
               </Typography>
 
-              {currentStage.floor != 1000 && (
+              {mentorInfo.currentStage.floor != 1000 && (
                 <StageProgress
-                  value={value}
-                  max={currentStage.max || 0}
-                  percent={currentStage.percent || 0}
+                  value={mentorInfo.value}
+                  max={mentorInfo.currentStage.max || 0}
+                  percent={mentorInfo.currentStage.percent || 0}
                 />
               )}
             </Grid>
             <Grid xs={12} md={2}>
               <RecommendedActionButton
                 mentor={props.mentor}
-                accessToken={props.accessToken}
                 setThumbnail={updateThumbnail}
               />
             </Grid>
@@ -321,9 +189,9 @@ export default function MyMentorCard(props: {
         </CardContent>
       </Card>
       <StageToast
-        value={value}
-        floor={currentStage.floor}
-        name={currentStage.name}
+        value={mentorInfo.value}
+        floor={mentorInfo.currentStage.floor}
+        name={mentorInfo.currentStage.name}
       />
     </div>
   );
